Build scenic_driver_nerves_rpi under _build
Move the build products out of the source tree and under _build. This
lets you build for different environments (dev, prod, test) and
different target hardware without cleaning.

This also copies the fonts to the _build directory. This avoids confusion
between the _build/.../priv and the source priv directories and who's
responsible for managing them (elixir_make or this Makefile).

Since nothing should be in the source priv directory, this commit also
removes it from the .gitignore.

Finally, the elixir_make dependency is bumped to "~> 0.6" to make sure
that the environment variables needed by the Makefile are set.